+++ common/src/main/java/de/cech12/usefulhats/Constants.java
@@ -2,7 +2,7 @@ package de.cech12.usefulhats;
 
 import de.cech12.usefulhats.item.EnderHelmetItem;
 import net.minecraft.core.component.DataComponentType;
-import net.minecraft.resources.ResourceLocation;
+import net.minecraft.resources.Identifier;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -29,8 +29,8 @@ public class Constants {
 
     private Constants() {}
 
-    public static ResourceLocation id(String name) {
-        return ResourceLocation.fromNamespaceAndPath(MOD_ID, name);
+    public static Identifier id(String name) {
+        return Identifier.fromNamespaceAndPath(MOD_ID, name);
     }
 
 }