+++ fabric/src/main/java/de/cech12/usefulhats/compat/ModMenuCompat.java
@@ -3,7 +3,7 @@ package de.cech12.usefulhats.compat;
 import com.terraformersmc.modmenu.api.ConfigScreenFactory;
 import com.terraformersmc.modmenu.api.ModMenuApi;
 import de.cech12.usefulhats.platform.FabricConfigHelper;
-import me.shedaniel.autoconfig.AutoConfig;
+import me.shedaniel.autoconfig.AutoConfigClient;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 
@@ -12,7 +12,7 @@ public class ModMenuCompat implements ModMenuApi {
 
     @Override
     public ConfigScreenFactory<?> getModConfigScreenFactory() {
-        return parent -> AutoConfig.getConfigScreen(FabricConfigHelper.class, parent).get();
+        return parent -> AutoConfigClient.getConfigScreen(FabricConfigHelper.class, parent).get();
     }
 
 }